requirements.txt
django
djangorestframework
pycryptodome
debugpy
kubernetes
simplejson
pyyaml
cerberus
pymysql
django-cors-headers
djangorestframework-simplejwt